Small changes for ne GUI
Added correct price scale for product purchase. Just a little more than if you purchase a pallet. The backup prices are to cheap or expensive so it needs to be set ;-)

Added the camera as it is an option in the GUI now,

If you do not want the small UI just change `showPopupUI` to false

Fixed some English translations. and Remove CDATA from 'title' it is not needed for modHub